maint: move pixeldata interpretation into core library, remove bytebuffer dependency, refactor the pixeldata buffers into interators over the interpreted pixels
