+++ dcmpipe_cli/Cargo.toml
@@ -11,7 +11,6 @@ default = ["index"]
 [dependencies]
 anyhow = "1.0"
 bson = { version = "2.13", optional = true }
-bytebuffer = "2.3"
 clap = { version = "4.5", features = ["derive"] }
 crossterm = "0.28"
 dcmpipe_lib = { path = "../dcmpipe_lib", version = "0.1", features = ["compress", "dimse", "stddicom"] }
